Code corrections and comments

+++ src/Ghost.cpp
@@ -12,9 +12,11 @@ Ghost::Ghost(int i_id)
     id = i_id;
 }
 
-
+// We check cells from all 4 directions.
 bool Ghost::pacman_collision(const Position& i_pacman_position)
 {
+    // We check collision by adding and subtracting one cell size from Pacmans position (We get a range).
+    // If ghost position is present in that range then collision is present.
     if(position.x > i_pacman_position.x - CELL_SIZE && position.x < i_pacman_position.x + CELL_SIZE)
     {
         if(position.y > i_pacman_position.y - CELL_SIZE && position.y < i_pacman_position.y + CELL_SIZE)
@@ -32,6 +34,7 @@ float Ghost::get_target_distance(int i_direction)
     int x = position.x;
     int y = position.y;
 
+    // We'll imaginarily move the gohst in a given direction and calculate the distance to the target.
     switch (i_direction)
     {
         case 0:
@@ -56,13 +59,16 @@ float Ghost::get_target_distance(int i_direction)
         }
     }
 
+    // Used Pythagoras' theorem to calculate the distance.
     return static_cast<float>(sqrt(pow(x - target.x, 2) + pow(y - target.y, 2)));
 }
 
 
 void Ghost::draw(bool i_flash, sf::RenderWindow& i_window)
 {
-    int body_frame = static_cast<int>(floor(animation_timer / static_cast<int>(GHOST_ANIMATION_SPEED)));
+    // To get the Current frame of the animation.
+    int body_frame = static_cast<int>(floor(animation_timer / static_cast<float>(GHOST_ANIMATION_SPEED)));
+    // We move to next frame when the timer value passes fixed time interval (ghost animation speed).
 
     sf::Sprite body;
     sf::Sprite face;
@@ -71,6 +77,8 @@ void Ghost::draw(bool i_flash, sf::RenderWindow& i_window)
     texture.loadFromFile("src/Resources/Images/Ghost" + std::to_string(CELL_SIZE) + ".png");
     body.setTexture(texture);
     body.setPosition(position.x, position.y);
+
+    // Animating ghost as per the animation frame.
     body.setTextureRect(sf::IntRect(CELL_SIZE * body_frame, 0, CELL_SIZE, CELL_SIZE));
 
     face.setTexture(texture);
@@ -111,7 +119,7 @@ void Ghost::draw(bool i_flash, sf::RenderWindow& i_window)
         body.setColor(sf::Color(36,36,255));
         face.setTextureRect(sf::IntRect(4 * CELL_SIZE, CELL_SIZE, CELL_SIZE, CELL_SIZE));
 
-        if(i_flash && body_frame % 2 == 0)
+        if(i_flash && body_frame % 2 == 0) // We change color after every 2 frames to get flash effect.
         {
             body.setColor(sf::Color(255,255,255));
             face.setColor(sf::Color(255,0,0));
@@ -125,6 +133,7 @@ void Ghost::draw(bool i_flash, sf::RenderWindow& i_window)
         i_window.draw(body);
     }
 
+    // Here we only draw face.
     else
     {
         face.setTextureRect(sf::IntRect(CELL_SIZE * direction, 2 * CELL_SIZE, CELL_SIZE, CELL_SIZE));
@@ -132,6 +141,7 @@ void Ghost::draw(bool i_flash, sf::RenderWindow& i_window)
 
     i_window.draw(face);
 
+    // To loop animation timer in fixed range (0, GHOST_ANIMATION_FRAMES * GHOST_ANIMATION_SPEED).
     animation_timer = (1 + animation_timer) % (GHOST_ANIMATION_FRAMES * GHOST_ANIMATION_SPEED);
 }
 
@@ -169,14 +179,16 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
 {
     bool move = 0;
 
+    // We check for all possible cells that the ghost can move to.
     int available_ways = 0;
     int speed = GHOST_SPEED;
 
     std::array<bool, 4> walls{};
 
+    // Here the gohst starts and stops being frightened.
     if(frightened_mode == 0 && i_pacman.get_energizer_timer() == ENERGIZER_DURATION / pow(2, i_level))
     {
-        frightened_speed_timer = GHOST_FRIGHTENED_SPEED;
+        frightened_speed_timer = GHOST_FRIGHTENED_SPEED; // To reduce ghost speed in frightened mode.
         frightened_mode = 1;
     }
     else if(i_pacman.get_energizer_timer() == 0 && frightened_mode == 1)
@@ -184,25 +196,32 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
         frightened_mode = 0;
     }
 
+    // In case the gohst goes outside the grid.
     if(frightened_mode == 2 && position.x % GHOST_ESCAPE_SPEED == 0 && position.y % GHOST_ESCAPE_SPEED == 0)
     {
         speed = GHOST_ESCAPE_SPEED;
     }
 
+    // To update the target position of the ghost.
     update_target(i_pacman.get_direction(), i_ghost_0.get_position(), i_pacman.get_position());
 
+    // We check for walls in all 4 directions.
     walls[0] = map_collision(0, use_door, speed + position.x, position.y, i_map);
     walls[1] = map_collision(0, use_door, position.x, position.y - speed, i_map);
     walls[2] = map_collision(0, use_door, position.x - speed, position.y, i_map);
     walls[3] = map_collision(0, use_door, position.x, speed + position.y, i_map);
 
+
+    // Here we set the direction of ghost.
     if(frightened_mode != 1)
     {
         int optimal_direction = 4;
         move = 1;
 
+        // We check for all possible cells that the ghost can move.
         for(int a=0; a < 4; a++)
         {
+            // Gohsts can't turn back (Unless they really have to).
             if(a == (direction + 2) % 4)
             {
                 continue;
@@ -218,6 +237,7 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
 
                 if(get_target_distance(a) < get_target_distance(optimal_direction))
                 {
+                    // The optimal direction is the direction that's closest to the target.
                     optimal_direction = a;
                 }
             }
@@ -225,10 +245,11 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
 
         if(available_ways > 1)
         {
+            // When the gohst is at the intersection, it chooses the optimal direction.
             direction = optimal_direction;
         }
         else
-        {
+        {   // If there's no other way, it turns back.
             if(optimal_direction == 4)
             {
                 direction = (direction + 2) % 4;
@@ -242,8 +263,10 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
 
     else
     {
+        // To give random movement to frightened ghost.
         int random_direction = rand() % 4;
 
+        // The gohst can move after a certain number of frames.
         if(frightened_speed_timer == 0)
         {
             move = 1;
@@ -265,12 +288,14 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
             {
                 while (walls[random_direction] == 1 || random_direction == (direction + 2) % 4)
                 {
+                    // We keep picking a random direction until we can use it.
                     random_direction = rand() % 4;
                 }
                 direction = random_direction;
             }
             else
             {
+                // If there's no other way, it turns back.
                 direction = (direction + 2) % 4;
             }
         }
@@ -280,6 +305,7 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
         }
     }
 
+    // Actual movement of ghost in optimal direction.
     if(move == 1)
     {
         switch (direction)
@@ -306,6 +332,7 @@ void Ghost::update(int i_level, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WID
             }
         }
 
+        // Warping.
         if(-CELL_SIZE >= position.x)
         {
             position.x = CELL_SIZE * MAP_WIDTH - speed;
@@ -351,8 +378,9 @@ void Ghost::update_target(int i_pacman_direction, const Position& i_ghost_0_posi
     }
     else
     {
-        if(movement_mode == 0)
+        if(movement_mode == 0) // The scatter mode.
         {
+            // Each gohst goes to the corner it's assigned to.
             switch (id)
             {
                 case 0:
@@ -377,17 +405,17 @@ void Ghost::update_target(int i_pacman_direction, const Position& i_ghost_0_posi
                 }
             }
         }
-        else
+        else  // The chase mode.
         {
             switch (id)
             {
-                case 0: // Red
+                case 0: // The red gohst will chase Pacman.
                 {
                     target = i_pacman_position; 
                     break;
                 }
 
-                case 1: // Pink
+                case 1: // The pink gohst will chase the 4th cell in front of Pacman.
                 {
                     target = i_pacman_position;
 
@@ -421,6 +449,7 @@ void Ghost::update_target(int i_pacman_direction, const Position& i_ghost_0_posi
                 {
                     target = i_pacman_position;
 
+                    // Getting the second cell in front of Pacman.
                     switch (i_pacman_direction)
                     {
                         case 0:
@@ -445,12 +474,14 @@ void Ghost::update_target(int i_pacman_direction, const Position& i_ghost_0_posi
                         }
                     }
 
+                    // We're sending a vector from the red gohst to the second cell in front of Pacman.
+					// Then we're doubling it.
                     target.x += target.x - i_ghost_0_position.x;
                     target.y += target.y - i_ghost_0_position.y;
                     break;
                 }
 
-                case 3: // Orange
+                case 3: // The orange gohst will chase Pacman until it gets close to him. Then it'll switch to the scatter mode.
                 {
                     if(CELL_SIZE * GHOST_3_CHASE <= sqrt(pow(position.x - i_pacman_position.x, 2) + pow(position.y - i_pacman_position.y, 2)))
                     {
